Extract form elements to their own helper
These are quite long of necessity, so that's overridden in Rubocop.
